telemetry: clean up test code (#27813)
