+++ telemetry/data_feature_usage_test.go
@@ -18,20 +18,16 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/telemetry"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/stretchr/testify/require"
-	"github.com/tikv/client-go/v2/testutils"
 )
 
 func TestTxnUsageInfo(t *testing.T) {
 	t.Parallel()
 
-	store, clean := newMockStore(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	t.Run("Used", func(t *testing.T) {
@@ -76,7 +72,7 @@ func TestTxnUsageInfo(t *testing.T) {
 func TestTemporaryTable(t *testing.T) {
 	t.Parallel()
 
-	store, clean := newMockStore(t)
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
@@ -91,25 +87,4 @@ func TestTemporaryTable(t *testing.T) {
 	usage, err = telemetry.GetFeatureUsage(tk.Session())
 	require.NoError(t, err)
 	require.True(t, usage.TemporaryTable)
-}
-
-func newMockStore(t *testing.T) (store kv.Storage, clean func()) {
-	store, err := mockstore.NewMockStore(
-		mockstore.WithClusterInspector(func(c testutils.Cluster) {
-			mockstore.BootstrapWithSingleStore(c)
-		}),
-		mockstore.WithStoreType(mockstore.EmbedUnistore),
-	)
-	require.NoError(t, err)
-
-	dom, err := session.BootstrapSession(store)
-	require.NoError(t, err)
-
-	clean = func() {
-		dom.Close()
-		err := store.Close()
-		require.NoError(t, err)
-	}
-
-	return
 }
