+++ telemetry/data_window_test.go
@@ -8,8 +8,10 @@
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 package telemetry_test
 
 import (
@@ -21,7 +23,9 @@ import (
 )
 
 func TestBuiltinFunctionsUsage(t *testing.T) {
-	store, clean := newMockStore(t)
+	t.Parallel()
+
+	store, clean := testkit.CreateMockStore(t)
 	defer clean()
 
 	tk := testkit.NewTestKit(t, store)
