+++ telemetry/telemetry_test.go
@@ -21,11 +21,9 @@ import (
 
 	"github.com/Jeffail/gabs/v2"
 	"github.com/pingcap/tidb/config"
-	"github.com/pingcap/tidb/domain"
-	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/session"
-	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/telemetry"
+	"github.com/pingcap/tidb/testkit"
 	"github.com/stretchr/testify/require"
 	"go.etcd.io/etcd/integration"
 )
@@ -37,18 +35,18 @@ func TestTrackingID(t *testing.T) {
 
 	t.Parallel()
 
-	s := newSuite(t)
-	defer s.close()
+	etcdCluster := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
+	defer etcdCluster.Terminate(t)
 
-	id, err := telemetry.GetTrackingID(s.etcdCluster.RandClient())
+	id, err := telemetry.GetTrackingID(etcdCluster.RandClient())
 	require.NoError(t, err)
 	require.Equal(t, "", id)
 
-	id2, err := telemetry.ResetTrackingID(s.etcdCluster.RandClient())
+	id2, err := telemetry.ResetTrackingID(etcdCluster.RandClient())
 	require.NoError(t, err)
 	require.NotEqual(t, "", id2)
 
-	id3, err := telemetry.GetTrackingID(s.etcdCluster.RandClient())
+	id3, err := telemetry.GetTrackingID(etcdCluster.RandClient())
 	require.NoError(t, err)
 	require.Equal(t, id2, id3)
 }
@@ -58,19 +56,24 @@ func TestPreview(t *testing.T) {
 		t.Skip("integration.NewClusterV3 will create file contains a colon which is not allowed on Windows")
 	}
 
-	s := newSuite(t)
-	defer s.close()
+	etcdCluster := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
+	defer etcdCluster.Terminate(t)
+	store, clean := testkit.CreateMockStore(t)
+	defer clean()
+	se, err := session.CreateSession4Test(store)
+	require.NoError(t, err)
+	defer se.Close()
 
 	config.GetGlobalConfig().EnableTelemetry = false
-	r, err := telemetry.PreviewUsageData(s.se, s.etcdCluster.RandClient())
+	r, err := telemetry.PreviewUsageData(se, etcdCluster.RandClient())
 	require.NoError(t, err)
 	require.Equal(t, "", r)
 
-	trackingID, err := telemetry.ResetTrackingID(s.etcdCluster.RandClient())
+	trackingID, err := telemetry.ResetTrackingID(etcdCluster.RandClient())
 	require.NoError(t, err)
 
 	config.GetGlobalConfig().EnableTelemetry = true
-	r, err = telemetry.PreviewUsageData(s.se, s.etcdCluster.RandClient())
+	r, err = telemetry.PreviewUsageData(se, etcdCluster.RandClient())
 	require.NoError(t, err)
 
 	jsonParsed, err := gabs.ParseJSON([]byte(r))
@@ -86,79 +89,17 @@ func TestPreview(t *testing.T) {
 	require.Equal(t, "tikv", jsonParsed.Path("instances.1.instanceType").Data().(string))
 	require.True(t, jsonParsed.ExistsP("hardware"))
 
-	_, err = s.se.Execute(context.Background(), "SET @@global.tidb_enable_telemetry = 0")
+	_, err = se.Execute(context.Background(), "SET @@global.tidb_enable_telemetry = 0")
 	require.NoError(t, err)
-	r, err = telemetry.PreviewUsageData(s.se, s.etcdCluster.RandClient())
+	r, err = telemetry.PreviewUsageData(se, etcdCluster.RandClient())
 	require.NoError(t, err)
 	require.Equal(t, "", r)
 
-	_, err = s.se.Execute(context.Background(), "SET @@global.tidb_enable_telemetry = 1")
+	_, err = se.Execute(context.Background(), "SET @@global.tidb_enable_telemetry = 1")
 	config.GetGlobalConfig().EnableTelemetry = false
 	require.NoError(t, err)
 
-	r, err = telemetry.PreviewUsageData(s.se, s.etcdCluster.RandClient())
+	r, err = telemetry.PreviewUsageData(se, etcdCluster.RandClient())
 	require.NoError(t, err)
 	require.Equal(t, "", r)
-}
-
-func TestReport(t *testing.T) {
-	if runtime.GOOS == "windows" {
-		t.Skip("integration.NewClusterV3 will create file contains a colon which is not allowed on Windows")
-	}
-
-	s := newSuite(t)
-	defer s.close()
-
-	config.GetGlobalConfig().EnableTelemetry = false
-	err := telemetry.ReportUsageData(s.se, s.etcdCluster.RandClient())
-	require.NoError(t, err)
-
-	status, err := telemetry.GetTelemetryStatus(s.etcdCluster.RandClient())
-	require.NoError(t, err)
-
-	jsonParsed, err := gabs.ParseJSON([]byte(status))
-	require.NoError(t, err)
-	require.True(t, jsonParsed.Path("is_error").Data().(bool))
-	require.Equal(t, "telemetry is disabled", jsonParsed.Path("error_msg").Data().(string))
-	require.False(t, jsonParsed.Path("is_request_sent").Data().(bool))
-}
-
-type testSuite struct {
-	store       kv.Storage
-	dom         *domain.Domain
-	etcdCluster *integration.ClusterV3
-	se          session.Session
-	close       func()
-}
-
-func newSuite(t *testing.T) *testSuite {
-	suite := new(testSuite)
-
-	store, err := mockstore.NewMockStore()
-	require.NoError(t, err)
-	suite.store = store
-
-	session.SetSchemaLease(0)
-	session.DisableStats4Test()
-
-	dom, err := session.BootstrapSession(store)
-	require.NoError(t, err)
-	suite.dom = dom
-
-	etcdCluster := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
-	suite.etcdCluster = etcdCluster
-
-	se, err := session.CreateSession4Test(store)
-	require.NoError(t, err)
-	suite.se = se
-
-	suite.close = func() {
-		suite.se.Close()
-		suite.etcdCluster.Terminate(t)
-		suite.dom.Close()
-		err = suite.store.Close()
-		require.NoError(t, err)
-	}
-
-	return suite
 }
